+++ examples/chain_client/exchange/query/17_SpotOrdersByHashes.py
@@ -26,7 +26,7 @@ async def main() -> None:
 
     subaccount_id = address.get_subaccount_id(index=0)
 
-    orders = await client.fetch_chain_spot_orders_by_hashes(
+    orders = await client.fetch_chain_spot_orders_by_hashes_v2(
         market_id="0x0611780ba69656949525013d947713300f56c37b6175e02f26bffa495c3208fe",
         subaccount_id=subaccount_id,
         order_hashes=["0x57a01cd26f1e2080860af3264e865d7c9c034a701e30946d01c1dc7a303cf2c1"],