+++ examples/chain_client/exchange/query/28_DerivativeMarket.py
@@ -11,7 +11,7 @@ async def main() -> None:
     # initialize grpc client
     client = AsyncClient(network)
 
-    derivative_market = await client.fetch_chain_derivative_market(
+    derivative_market = await client.fetch_chain_derivative_market_v2(
         market_id="0x17ef48032cb24375ba7c2e39f384e56433bcab20cbee9a7357e4cba2eb00abe6",
     )
     print(derivative_market)