+++ examples/chain_client/exchange/query/49_HistoricalTradeRecords.py
@@ -11,7 +11,7 @@ async def main() -> None:
     # initialize grpc client
     client = AsyncClient(network)
 
-    records = await client.fetch_historical_trade_records(
+    records = await client.fetch_historical_trade_records_v2(
         market_id="0x0611780ba69656949525013d947713300f56c37b6175e02f26bffa495c3208fe"
     )
     print(records)