+++ examples/chain_client/exchange/query/57_GrantAuthorization.py
@@ -1,0 +1,34 @@
+import asyncio
+import os
+
+import dotenv
+
+from pyinjective import PrivateKey
+from pyinjective.async_client import AsyncClient
+from pyinjective.core.network import Network
+
+
+async def main() -> None:
+    dotenv.load_dotenv()
+    configured_private_key = os.getenv("INJECTIVE_PRIVATE_KEY")
+    grantee_public_address = os.getenv("INJECTIVE_GRANTEE_PUBLIC_ADDRESS")
+
+    # select network: local, testnet, mainnet
+    network = Network.testnet()
+
+    # initialize grpc client
+    client = AsyncClient(network)
+
+    priv_key = PrivateKey.from_hex(configured_private_key)
+    pub_key = priv_key.to_public_key()
+    address = pub_key.to_address()
+
+    active_grant = await client.fetch_grant_authorization(
+        granter=address.to_acc_bech32(),
+        grantee=grantee_public_address,
+    )
+    print(active_grant)
+
+
+if __name__ == "__main__":
+    asyncio.get_event_loop().run_until_complete(main())