+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.9.0] - 9999-99-99
+### Added
+- Added support for Exchange V2 proto queries and types
+- Updated all chain exchange module examples to use the new Exchange V2 proto queries and types
+
+### Removed
+- Removed all methods marked as deprecated in AsyncClient and Composer
+
 ## [1.8.0] - 2024-11-14
 ### Changed
 - The markets initialization in AsyncClient has been modified to get markets information from the chain endpoints instead of the Indexer endpoints